pyproject.toml
[project]
name = "beacon-fever-differential"
version = "1.0.0"
description = "BEACON: Bayesian Evidence-Adjusted Clinical Outbreak Navigator — Australian-weighted diagnostic differential for febrile returned travellers"
readme = "README.md"
license = "CC-BY-4.0"
requires-python = ">=3.10"
authors = [
    { name = "Hayden Farquhar", email = "hayden.farquhar@icloud.com" },
]
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "feedparser>=6.0",
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
    "tabula-py>=2.9",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "streamlit>=1.32",
    "altair>=5.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
